+++ lib/src/configs/tree_rename_selection.dart
@@ -1,0 +1,2 @@
+export 'package:super_tree/src/configs/rename_selection/tree_rename_selection_core.dart';
+export 'package:super_tree/src/configs/rename_selection/tree_rename_selection_strategies.dart';